@@ -4,14 +4,19 @@
     "license": "MIT",
     "type": "library",
     "require": {
-        "php": ">=8.1",
-        "psr/http-client": "^1.0"
+        "php": ">=8.2",
+        "psr/http-client": "^1.0",
+        "ext-openssl": "*",
+        "symfony/validator": "^6.3|^5.4",
+        "symfony/serializer": "^6.3|^5.4",
+        "symfony/property-access": "^6.3|^5.4"
     },
     "require-dev": {
-        "symfony/http-client": "^5.1",
-        "symfony/var-dumper": "^5.1",
+        "symfony/http-client": "^6.3",
+        "symfony/var-dumper": "^v6.3",
         "nyholm/psr7": "^1.3",
-        "phpunit/phpunit": "^9"
+        "phpunit/phpunit": "^9",
+        "symfony/dependency-injection": "^6.3"
     },
     "autoload": {
         "psr-4": {
